Kosten baten afschrijvingen (#730)
Cost benefit: add depreciation costs "afschrijvingen" to table

Each (depreciation) costs query in the costs config that has datamodel assets attached through a datamodel query gets an extra filter to calculate the split in assets between actors. This split is used to add depreciation costs for each actor to the costs table. Unregocnized costs are sent to an Undefined post.

Co-authored-by: ALFP <pascal.alferink@witteveenbos.com>
Co-authored-by: Erik van Velzen <erik@zenmo.com>