Interactive element: allow relations
Enable filtering and change relations by id with interactive elements.
These options were removed with jQuery.

This is now only allowed for actor group and subgroup due to
existing backend validation.

The ideal situation would be to make the front-end options dependent on the
backend validation so it is always in sync (single source of truth).